Error handling, retries and faster tuned reloads.
This PR improves error handling and adds a simple retry mechanism for the main
loop.  This mostly fixes situations when the pod watch channel is closed.

Node watch was removed as it is expensive and "fires" far too often when node
labels are *not* actually changed.

A workaround to improve the tuned reload speed (lower the soft limit of open
file descriptors) is introduced.