Support for the "rendered" Tuned CR.
Recent NTO stops using ConfigMaps to communicate profile changes between
the operator and this operands.  This change adds support for the "rendered"
Tuned CR which is now used by the operator to inform the operand about
the global tuned profile changes.  Support for the old way of communicating
the changes (ConfigMaps) is still present, but will be removed soon.